requirements.txt
pyyaml>=6.0

# Cloud extras (install only if needed)
# google-cloud-storage>=2.10
# google-cloud-aiplatform>=1.50
